bridge config update
