+++ backend/cmd/bridge/config/colors.go
