+++ backend/cmd/bridge/config/generator.go
