+++ backend/cmd/bridge/config/validator.go
